+++ src/alt/utils/AltUtils.js
@@ -18,9 +18,11 @@ export function getInternalMethods(Obj, isProto) {
 
 export function warn(msg) {
   /* istanbul ignore else */
+  /*eslint-disable*/
   if (typeof console !== 'undefined') {
     console.warn(new ReferenceError(msg))
   }
+  /*eslint-enable*/
 }
 
 export function uid(container, name) {