+++ src/utils/AltIso.js
@@ -8,21 +8,21 @@ export default {
     if (typeof window === 'undefined') {
       return Render.toString(alt, Component, props).then((obj) => {
         return {
-          html: Iso.render(obj.html, obj.state, { iso: 1 })
+          html: Iso.render(obj.html, obj.state, { iso: 1 }),
         }
       }).catch((err) => {
         // return the empty markup in html when there's an error
         return {
           err,
-          html: Iso.render()
+          html: Iso.render(),
         }
       })
-    } else {
-      Iso.bootstrap((state, meta, node) => {
-        alt.bootstrap(state)
-        Render.toDOM(Component, props, node, meta.iso)
-      })
-      return Promise.resolve()
     }
-  }
+
+    Iso.bootstrap((state, meta, node) => {
+      alt.bootstrap(state)
+      Render.toDOM(Component, props, node, meta.iso)
+    })
+    return Promise.resolve()
+  },
 }